tines: Building the Prongs of Your Forking Paths

Tools for documenting analytical decisions in data analysis. Analytical decisions are documented as a schema of steps, and large language models are used to generate and validate R code. Schemas can be expanded with alternative decisions into a multiverse of analyses, allowing analysts to explore how their conclusions depend on the analytical choices they made.

See also

Author

Maintainer: H. Sherry Zhang huizezhangsh@gmail.com [copyright holder]